backend/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Chat app server"
authors = [
    {name = "Anas EL MOUDEN"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.118.0,<0.119.0)",
    "uvicorn[standard] (>=0.37.0,<0.38.0)",
    "sqlmodel (>=0.0.25,<0.0.26)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic[email] (>=2.11.9,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "requests>=2.32.5",
]

[build-system]
build-backend = "hatchling.build"
requires = [
    "hatchling",
    "versioningit>=3.1.2",
]
